+++ packages/go_router/lib/go_router.dart
@@ -9,7 +9,6 @@ library go_router;
 export 'src/configuration.dart' show GoRouterState, GoRoute;
 export 'src/misc/extensions.dart';
 export 'src/misc/inherited_router.dart';
-export 'src/misc/refresh_stream.dart';
 export 'src/pages/custom_transition_page.dart';
 export 'src/platform.dart' show UrlPathStrategy;
 export 'src/route_data.dart' show GoRouteData, TypedGoRoute;